src/main/kotlin/blackjack/domain/state/Hit.kt
package blackjack.domain.state

import blackjack.domain.card.Card
import blackjack.domain.card.Cards

class Hit(cards: Cards) : State(cards) {
    override fun draw(card: Card): State {
        val newCards: Cards = cards + card
        return when {
            newCards.calculateScore().isBlackJack -> BlackJack(newCards)
            newCards.calculateScore().isBust -> Bust(newCards)
            else -> Hit(cards + card)
        }
    }

    override fun matchWith(otherState: State): Outcome = Stay(cards).matchWith(otherState)
}
